+++ test/_env/util.js
@@ -87,6 +87,7 @@ function splitMultipartResponse(body, boundary = "boundary") {
         const subBoundary = multipart[1];
         return splitMultipartResponse(_rest.join("\r\n\r\n"), subBoundary);
       } else {
+        const contentID = _meta.match(/content-id:\s*(\w+)/i);
         const [_info, _body] = _rest;
         const body = _body && _body.startsWith("{") ? JSON.parse(_body) : _body;
         const [_status, ..._headers] = _info.split("\r\n");
@@ -97,7 +98,7 @@ function splitMultipartResponse(body, boundary = "boundary") {
           const [key, value] = _header.split(": ");
           headers[key] = value;
         });
-        return { statusCode, statusText, headers, body };
+        return { statusCode, statusText, headers, body, contentID: contentID && contentID[1] };
       }
     });
 }